UI-UX updates
