+++ ios/Sentiment Bot/Models/ErrorMessage.swift
@@ -8,7 +8,6 @@
 
 import Foundation
 
-
 struct ErrorMessage: Codable {
     let message: [String]
 }
