+++ ios/Sentiment Bot/View Controllers/MembersTableViewCell.swift
@@ -13,11 +13,5 @@ class MembersTableViewCell: UITableViewCell {
     @IBOutlet weak var memberImageView: UIImageView!
     @IBOutlet weak var nameLabel: UILabel!
 
-    func setResponse(response: Response) {
-//        nameLabel.text = "\(response.firstName) \(response.lastName)"
-//        cell?.imageView?.layer.cornerRadius = 45 //(cell?.imageView?.frame.size.width)! / 2
-//        cell?.imageView?.clipsToBounds = true
-//        cell?.imageView?.image = #imageLiteral(resourceName: "missing-image-5")
-    }
 
 }
